+++ tic.R
@@ -21,18 +21,4 @@ if (ci_get_env("TEST") == "Learner") {
       filter = paste0("_", tic::ci_get_env("PKG"), "_"),
       stop_on_failure = TRUE))
 
-} else {
-  do_package_checks()
-
-  #get_stage("before_deploy") %>%
-  #  add_step(step_setup_push_deploy(branch = "main", orphan = FALSE))
-  
-  get_stage("install") %>%
-    add_step(step_install_github("mlr-org/mlr3pkgdowntemplate"))
-
-  do_pkgdown()
-
- # get_stage("deploy") %>%
- #   add_code_step(rmarkdown::render("README.Rmd")) %>%
- #   add_step(step_do_push_deploy(commit_paths = c("README.md")))
 }