+++ NEWS.md
@@ -2,6 +2,14 @@
 
 * Added `LearnerRegrGam` and `LearnerClassifGam` with keys `regr.gam` and `classif.gam` from package `mgcv`.
 
+# mlr3extralearners 0.3.6
+
+* `surv.coxboost` now uses the GitHub version instead of CRAN (archived)
+
+# mlr3extralearners 0.3.4
+
+* Add support for custom families to `regr.glmboost`
+
 # mlr3extralearners 0.3.1
 
 * `surv.svm` now supports all feature types